# docs/swagger.yaml
definitions:
  auth.LoginRequestDto:
    properties:
      email:
        maxLength: 254
        type: string
      senha:
        maxLength: 72
        minLength: 8
        type: string
    required:
    - email
    - senha
    type: object
  auth.SessaoResponseDto:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
      token_type:
        type: string
      usuario:
        $ref: '#/definitions/auth.UsuarioResumo'
    type: object
  auth.TokenRequestDto:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  auth.UsuarioResumo:
    properties:
      email:
        type: string
      nome:
        type: string
      uuid:
        type: string
    type: object
...
  rest_err.RestErr:
    properties:
      code:
        type: string
      error:
        type: string
      message:
        type: string
      ray_trace:
        type: string
    type: object
...
info:
  contact: {}
paths:
  /api/application/identidade/auth/login:
    post:
      consumes:
      - application/json
      description: 'Login com e-mail/senha contra a organization resolvida pelo Host
        (subdomínio de workspace ou domínio custom). Falhas são indistinguíveis: usuário
        inexistente, senha errada e host sem organization devolvem o mesmo 401 genérico'
      parameters:
      - description: Credenciais
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.LoginRequestDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.SessaoResponseDto'
        "400":
          description: Corpo malformado
          schema:
            $ref: '#/definitions/rest_err.RestErr'
        "401":
          description: Credenciais inválidas (genérico)
          schema:
            $ref: '#/definitions/rest_err.RestErr'
      summary: Autentica e abre uma sessão
      tags:
      - Identidade · Auth
  /api/application/identidade/auth/logout:
    post:
      consumes:
      - application/json
      description: Revoga o refresh token no Postgres (marca revogado_em); a linha
        do jti nunca é removida
      parameters:
      - description: Refresh token a revogar
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.TokenRequestDto'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Corpo malformado
          schema:
            $ref: '#/definitions/rest_err.RestErr'
        "401":
          description: Sessão inválida ou expirada
          schema:
            $ref: '#/definitions/rest_err.RestErr'
      summary: Encerra a sessão
      tags:
      - Identidade · Auth
  /api/application/identidade/auth/refresh:
    post:
      consumes:
      - application/json
      description: Troca um refresh token válido por um novo par de tokens; o jti
        precisa seguir ativo no Postgres (revogação persistida)
      parameters:
      - description: Refresh token vigente
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.TokenRequestDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.SessaoResponseDto'
        "400":
          description: Corpo malformado
          schema:
            $ref: '#/definitions/rest_err.RestErr'
        "401":
          description: Sessão inválida ou expirada
          schema:
            $ref: '#/definitions/rest_err.RestErr'
      summary: Renova a sessão
      tags:
      - Identidade · Auth
...
swagger: "2.0"